requirements.txt
streamlit>=1.31.0
pdfplumber>=0.10.3
PyMuPDF>=1.23.8
python-docx>=1.1.0
markdown>=3.5.2
litellm>=1.20.0
python-dotenv>=1.0.1
python-dotenv
